fixed a few type errors and added sync react hook
